Check for single values
Check for single values in a block, row, or column.
Made the functions a lot more generalized by adding the brc option.
This means that only one check function needs to be used, but that function
behaves differently depending on the brc value. The brc value determines
wether it checks on a block, row, or column.

Also modified the printfield function. Thie printing of horizontal lines,
and the row and column numbers should make it easier to debug what's going
on.