Shortened the check_num_brc* functions
Only one if-statement to maintain. That should make it easier.
Also seperated the loose from the normal function.
Cleaned up some extra linebreaks etc.
And ofcourse added a lot more text to the header file. This will
give the readers at home a lot more understanding of what we do.