Issue #1 Fixed where, which confusion.
We're used to go from i to a BRC in the solving. In the bruteforcing
we should go from BRC to i, not from BRC to where.
The variables that point to the same in the field are now called the
same, regardless of in which part they are invoked.