Added automatic bruteforcer
Besides manually bruteforcing the sudoku, we could leave it to the
computer to do that.

Here is, in short, how it works.
Find the first field that has no value. Set the value to the first
possibility. Add that value to the bftry bitfield (also in the
backup). Try to solve it. If that makes it invalid, restore the backup
and try the next value until all possibilities for that field have been
tried. Move on to the next field. Repeat until sudoku is solved.